+++ plugins/main/public/components/common/modules/buttons/generate_report.tsx
@@ -12,9 +12,7 @@
 
 import React from 'react';
 import { useAsyncAction } from '../../hooks';
-import { getUiSettings } from '../../../../kibana-services';
 import { ReportingService } from '../../../../react-services';
-import $ from 'jquery';
 import { WzButton } from '../../../common/buttons';
 import { connect } from 'react-redux';
 
@@ -31,32 +29,7 @@ export const ButtonModuleGenerateReport = connect(mapStateToProps)(
     ].every(Boolean);
     const totalResults = dataSourceSearchContext?.totalResults;
     const action = useAsyncAction(async () => {
-      const reportingService = new ReportingService();
-      const isDarkModeTheme = getUiSettings().get('theme:darkMode');
-      if (isDarkModeTheme) {
-        //Patch to fix white text in dark-mode pdf reports
-        const defaultTextColor = '#DFE5EF';
-
-        //Patch to fix dark backgrounds in visualizations dark-mode pdf reports
-        const $labels = $(
-          '.euiButtonEmpty__text, .echLegendItem, div.mtrVis__value ~ div',
-        );
-        const $vizBackground = $('.echChartBackground');
-        const defaultVizBackground = $vizBackground.css('background-color');
-
-        try {
-          $labels.css('color', 'black');
-          $vizBackground.css('background-color', 'transparent');
-          await reportingService.startVis2Png(moduleID, agent?.id || false);
-          $vizBackground.css('background-color', defaultVizBackground);
-          $labels.css('color', defaultTextColor);
-        } catch (e) {
-          $labels.css('color', defaultTextColor);
-          $vizBackground.css('background-color', defaultVizBackground);
-        }
-      } else {
-        await reportingService.startVis2Png(moduleID, agent?.id || false);
-      }
+      await new ReportingService()?.generateInContextPDFReport();
     }, [agent]);
 
     return (